made changes to the Profile pages
